apply HDR tone mapping, improve LUT, modify gamma correction
LDR 색상 공간에서 HDR 색상 공간으로 변경한 이후, ACES 톤 매핑을 구현하였다.
톤 매핑 구현에 따라 LUT가 포토샵 LUT 기능을 그대로 구현할 수 있게 향상되었다.
모든 픽셀 쉐이더에서 감마 디코딩을 수행하도록 변경하였으며 LUT 컴퓨트 쉐이더에서만 감마 인코딩을 한 번만 수행하도록 변경하였다.
필터 옵션을 구현하여 다양한 필터 옵션을 끄고 킬 수 있도록 변경하였다.
톤 매핑은 LUT 필터를 통해 진행할 수 있으며 현재 모든 필터가 켜져 있어도 실행이 가능하다.

+++ Client/X-MACHINA/src/RootSignature.cpp
@@ -93,7 +93,7 @@ RComPtr<ID3D12RootSignature> GraphicsRootSignature::Create()
 	D3D12_ROOT_SIGNATURE_DESC rootSignatureDesc{};
 	rootSignatureDesc.NumParameters     = (UINT)mParams.size();
 	rootSignatureDesc.pParameters       = mParams.data();
-	rootSignatureDesc.NumStaticSamplers = staticSamplers.size();
+	rootSignatureDesc.NumStaticSamplers = (UINT)staticSamplers.size();
 	rootSignatureDesc.pStaticSamplers	= staticSamplers.data();
 	rootSignatureDesc.Flags             = rootSignatureFlags;
